Fix the support for Ledger HW1/Nano legacy hardware
Ledger HW1/Nano have support for Bitcoin Cash in version 1.0.4, so check for
that. The call to getWalletPublicKey with cashAddr=True causes the legacy
devices to return an error we can use to detect it. The warning about cashddr
is suppressed as they have no display anyway. The 2FA dialog has been fixed by
converting the address to a legacy address.