Parse general P2SH in parse_scriptSig, serialize from deserialized
Add parsing of generic P2SH contract addresses to the transaction code.
Note that this still needs a check (maybe?) whether the previous transaction
actually had a P2SH output.

It is important to notice that a transaction's TXID is recalculated from
these values. As unknown scriptSigs need to still produce the right serialization,
the serialization code has likewise been changed to always use an input's scriptSig,
if available.