wallet: Infrastructure for forfeits
This adds support for handling forfeit addresses to the wallet.

Two new class members are introduced:

forfeit_map: This maps forfeit addresses to their underlying P2PKH
address (for which the public key also matches the forfeit address if
using the correct P2SH script for scriptSig generation).

forfeit_scripts:
This maps forfeit addresses to their corresponding P2SH scripts that
allow to spend the forfeit addresses as normal P2PKH (or by two sigs
in case a miner detects a double-spend).

Both of these objects are persisted to the wallet JSON file for easier
handling and to cache the transaction parsing results.

It also adds support to the various accessor function to more properly
represent forfeit addresses.

It still misses the transaction generating code as well as the code to
extract forfeits from incoming transactions that match the template
forfeit P2SH code. This will come in an upcoming commit.