wallet: Support for making transactions with forfeits
If the configuration variable use_forfeits is set to true, this
will generate transaction that have a "ZCF\x00" OP_RETURN marker
indicating the use of forfeits as well as having the special
forfeit smart contract output to a corresponding P2SH address.

The P2SH address is deterministically generated from the transaction's
inputs and the change address that is alsoused for the "forfeit
change". This means that it is possible to recreate the P2SH contents
just by trying reconstruction of all available outputs and checking
for a matching P2SH hash.