Added some more matches of bad server responses.
If it looks like a URL or HTML appears in the server response, then
print a critical warning and pop up the network dialog, asking the user
to switch servers.

If it looks like a name.tld appears in the response, suppress the error
message (real server responses never have WORD.WORD).  This way phishers
can't pass on any information to users that contain anything that even
remotely resembles a URL.

If the error message is otherwise kosher, forward it to the user.