Anti-Phishing fix -- transform server messages into GUI messages.
This is a different take on the anti-phishing fixup.  In this commit, we
translate known-good server replies for errors (which were obtained from
reading the BU and Bitcoin ABC sourcecode) into translated GUI-friendly
error messages.

If the groking fails, we turn the responses into a generic "Broadcast
failed" style message