Performance Tweak
I realized that re-computing the "whitelisted hostmap" is wasteful of
CPU resources since it rarely changes.  Instead we cache it each time a
server is added/removed from the whitelist.

Other functions such as pick_random_server() expect a hostmap (not a
host set) -- which is why this hostmap is needed in the first place.