initial commit to fix the "wallets stays open" bug
On window close the wallet stays open and continues to synchronize
needlessly.  This is not expected UX.

This has now been addressed in this commit.  Basically we have to keep
track and be sure not to send out defunct subscriptions when we switch
servers.

Also, the verifier and synchronizer are given a "cleaned_up" flag to
ignore latent/delayed callbacks on wallet close.  Closed means CLOSED!
:)