[Lib][All Gui] 'updated' callback -> 'wallet_updated' , 'blockchain_updated'
This is a much needed change.  The 'updated' callback was overloaded to
indicate the wallet history was updated as well as the blockchain was
updated.

This was too large a granularity for what we need.

Since the Qt gui can have multiple wallet windows open, this led to a
situation where the GUI wallet windows were updating themselves multiple
times per wallet that got a new history, even if the update didn't
concern their particular wallet.

So -- we split up the ambiguous 'updated' event into 'blockchain_updated'
and 'wallet_updated'.

blockchain_updated -- callbacks receive 2 ints a parameters (local_height,
server_height).  Usually they are equal but local_height may be larger
than server_height if the server we are connected to is lagging behind
other servers in terms of the latest blockchain blocks it's seen.

wallet_updated -- callbacks receive the 'wallet' instance as a
parameter. Client code can look at this instance parameter and ignore
the notification if the update was for a different wallet.

The old 'updated' event will still be sent if callbacks have asked for it --
but the code will issue deprecated warnings using print_error each time
this happens.

Old plugins will continue to function as will legacy code, but with
warnings.

This change will allow client code to better be informed about the
state changes of the wallet vs/ blockchain.

FWIW -- Electrum has done this also long ago.

Also changed:

the 'status' event now also sends a string 'status' as a parameter
(previously it sent no parameters).

I grep'd through all the code and fixed everything to work with the new
API.

Note that Android still is using the old 'updated' scheme and thus it
needs to be migrated over else it will issue deprecated warnings.