Fixup: If cashacct tx verif. fails due to unknown txid, handle it better
We must assume the lookup server is wrong and the EX server we are
connected to is correct.  So we wipe the tx from our caches. it can get
added again later if user tries to lookup the same cashacct tx's again.

This is to prevent the situation where testnet wallets looking up tx's
on mainnet's lookup server get tons of spammy errors from ElectrumX.