Made cashacct verifiction more resilient/less buggy on bad txs
If lookup server gives us tx's that don't verify, we bail earlier on a
block and also deal with the situation better.

We tell the UI also about the situation immediately, to avoid blocking
the GUI.

This can happen when on testnet in particular, since the lookup server
is returning mainnet blockchain data and testnet clients freak out.

Now we deal with it better.