Qt GUI Fix: On Python 3.7.3, PyQt5 5.12.3+, signals not cleaned up
In particular the cashaddr_toggled_signal was connected to objects that
were since deleted. According to Qt (and PyQt) documentation the
connections should get auto-reaped.  But Axel Gember was getting
"Wrapped C++ object deleted" exceptions related to this signal still
firing on deleted slots.

As a safety measure, we are auto-disconnecting this signal during object
cleanup for objects that connect to the global gui_object.cashaddr_toggled_signal