Windows Qt: Added preferences option for enabling/disabling FreeType
The default to this option is True for Windows < Win10, False otherwise.

The preferences option is only presented on Windows (sys.platform ==
win32 or cygwin).  On other platforms it does not appear and its config
key has no effect.

This addresses the fact that for some users FreeType fonts are a good
idea, but for others perhaps they would like finer grained control.

FreeType in general is good but it does lack subpixel rendering so for
some users the tradeoff there may not warrant using it.