Qt: Fixup for the preferenes dialog crash
The crash was only happening when scanning the QR cameras. If the camera
scan code was removed, the crash would stop happening.

I suspect that QtMultimedia doesn't like to be invoked from a button
clicked handler or an action handler, or while constructing the dialog,
or both.

So the fix is to do the camera scan from the top-level event loop via a
QueuedConnection.

We fire off a signal on showEvent of the preferences dialog, which is
connected via QueuedConnection to a handler that then imports
QtMultimedia and then does all population of the video device combobox.

This fix makes it so that I can no longer reproduce the crash.

The ultimate root of the problem likely is some but in PyQt5, as it does
not happen with PySide2.