Payment Requests: Added BitPay 2.0 JSON API
This replaces BIP70 for "*.bitpay.com" on https *only*.

Apparently BitPay prefers we use this different API for payment
requests. It still presents us with some challenges in that it relies on
PGP ultimately to verify sigs.

For now we disable their bizarre key verification scheme and just
rely on SSL verification which is very secure against MITM.

I am waiting to hear back from BitPay if it's ok to hard-code bitcoin
keys in our app, since we'd rather not rely on PGP.

But for now, SSL verification is very good, so we rely on that for
BitPay 2.0.