Linux: Resolve monospace font to the actual font family
Just using monospace isn't respecting the users font choice
so we need to resolve an actual family to support that.

This is not done in util.py as that gets loaded before Qt is
initialized. Note that some plugins can be loaded before the
Qt GUI is loaded and will not get the updated monospace font.