Qt GUI: Sign message warning text nit
The 'msg_sign' variable was class-level and was untranslated as a result
(language & i19n isn't set necessarily when the class is being evaluated by
Python). Moved this variable init insite the method implementation.
There was no good reason for it to be class-level since it was a very
rarely used variable.

Also made the sign_message callpath for multisig abort, not try and sign
the message with a bogus private key.