Added miminal slp.WalletData implementation
This implementation does the bare minimum of maintainting fast data
structures and keeps track of the things we need right now for SLP
awareness to prevent tokens.

We give this provisional data layout slp_data_version '3.9' which should
cause EC-SLP to rebuild slp data when switching between this and EC-SLP.

In the future we will add more structure to this data when we enable
validation.

Regardless -- we need to do away with the EC-SLP data layout as it is
somewhat inefficient.