Qt: Add a minimize to system tray option
When enabled, this hides the window when the user minimizes it. We need to
make sure that the window is not minimized anymore in bring_to_top, which
is a good idea in any case. This also adds a Show/Hide menu item to the
system tray menu which does the same as double clicking the icon.