[interface] send requests much faster
The limit of 100 was interacting poorly with latency. For example
if there was RTT of 300ms, then we could only do one request per
3 ms on average. For a fat Fusion wallet that needs to do thousands
of subscriptions on first connect, this means 3 seconds for every
thousand addresses, and so when we go above >6000 addresses we can
hit the 20 second "blockchain request timed out" problem.

This patch drops the latency effect by a factor of 20. We could
actually remove the 2000 unanswered_requests limit entirely but
I'm thinking it is worth keeping in just as a sanity check,
for now.

With this patch I can get ~1000 requests/sec throughput without
much issue, even going over Tor.