Disable cointext: support from send tab & remove cointext code
Due to the announcement by cointext that they will be ceasing operations
as of Nov. 14th, 2020, we need to disable this facility from Electron
Cash.  See announcement:

https://cointext.io/en/nov14/?utm_source=CoinText&utm_campaign=53b0c5ed7d-EMAIL_CAMPAIGN_2020_10_19_02_15

All code that was added for PR #1407 (and subsequent clean-up commits) has
been revered by this commit.  It's as if cointext: was never in the
codebase at all.