optimize coin consolidation
Testing the tool on a 40000+ UTXO wallet showed that it would not work without optimizing it.
The main optimization is to not read the transaction history from the wallet file once per transaction (0.5 s), but just reading it once in the constructor. Some other data is also cached and improves the overal performance, at the cost of some additional verbosity and code duplication.

I have not tested signing the transactions in a very large wallet, as my test wallet is watchonly.