ensure the thread is stopped before destroying or restarting it
The user must not be allowed to start the thread before it is finished (by going back and forth between wizard pages) and must not be allowed to close the wizard before the thread is finished (risk of core dump when garbage collecting a running thread).
To make this less painful, I need to figure out a way to stop the tx building algorithm in between two tx.