fix signing multiple transactions
When the coin consolidation generated more than one transaction, using the `ElectrumWindow.sign_tx` helper in a loop caused a crash due to QThread being destroyed while still running.
The way this helper creates a new QDialog with a QThread for each transaction seems to work only when there is a single transaction. Rather than trying to fix it, it is simpler and cleaner to just call `wallet.sign_tx` even if this requires dealing with the password prompt.