refactor the ConsolidateWorker's interruption flag
Add a method to request interruption, so that the caller does not need to deal with the mutex or with the boolean attribute. This also adds documentation about the workers behavior (signals emitted) when interrupted.
Give the mutex a more specific name.
Improve the way the boolean value is returned (avoid doing `if True: return True; return False`)