fix loading multiple signed tx from file
Raw signed transactions don't specify input amounts, so fees and amounts cannot be computed. This commit fixes the issue by showing "not available" for missing data.
A better solution would be to attempt to fetch data from the network, which would need to be done asynchronously because the typical use case for "Load multiple tx..." is to deal with many large consolidation transactions.

# Conflicts:
#	electroncash_gui/qt/multi_transactions_dialog.py