dynamicaly update the max value for the min amount widget
It does not make sense for the minimum amount to be larger than the maximum amount, so connect the `valueChanged` signal for the max amount widget to a method that updates the range of the minimum amount widget.

Rename the widgets to use the word `amount` instead of `value`, to reduce the confusiont. "Amount" now refers to the range defined by the two widgets' current values, and "value" is used by Qt to define the data in the individual widget (which happens to be an amount).

# Conflicts:
#	electroncash_gui/qt/consolidate_coins_dialog.py