don't change min spinbox range when max spinbox value is changed
The previous behavior caused issues when the users typed in a max value, because when typing the first few digits the value is temporarily lower than the value in the min spinbox. Setting a low max value on a spinbox than its current value causes it to overwrite the current value. So changing the max amount/height after changing the min amoun/height causes the min to be overwritten.

Instead of making sure the min spinbox cannot have a larger value than the max spinbox, just show the values in red when the range is inconsistent.

# Conflicts:
#	electroncash_gui/qt/consolidate_coins_dialog.py