Build: Fix PE file timestamp of MSVC generated binaries
These files included a `TimeDateStamp` in `IMAGE_NT_HEADERS32` and
`IMAGE_DEBUG_DIRECTORY` making the build not reproducible.

There is a `/Brepro` argument that can be passed to the linker to set
these to 0xffffffff but `setuptools` does not allow passing custom flags
in the environment for MSVC.

To fix this we include a small Python tool that uses the pefile library
to set the timestamps to 0xffffffff.