[qt] Display history in topological order per block
As discussed in telegram, it is confusing to users how the blockchain
ordering is used for the txns in the history tab. This is CTOR ordering
and sometimes it leads to intermediate balance values that make no
sense (spends coming before receives, negative temporary balances, etc).
This is because blockchain ordering is by txid and is not natural ordering
since the HF in Nov 2018.

Now, by default, the history will try and sort txs by "natural" order or
topologically. This will guarantee that receives of a coin come before
spends of the same coin in the history, thus the balance always is
positive and never dips into negatives, and is the natural order that
the txns likely were issued in.

A config option was added in the settings UI to revert to the old
behavior, but the default now is to sort in natural order rather than
the less useful CTOR ordering.

Special thanks to molecular123 and matricz for reminding me of this
issue.