[qt] Add option to history tab to always display receives before sends
This option is default-on.  This implementation is an alternative to
PR #2431.  It is a second attempt to "undo the damage" of CTOR on the
wallet UI and avoid temporary negative intermediate balances.

The approach taken here is to always sort txns in the same block by
receives-first.  Thus all txns for the same block that have a net
positive effect on wallet balance will appear before these txns in the
same block with a net negative effect on wallet balance.  This is a
quick and performant way to avoid intermediate negative wallet balances
in the history tab (due to CTOR).