Interface: Use check_hostname instead of match_hostname
Since Python 3.7 the hostname verification is done in
`SSLSocket.do_handshake` if `SSLContext.check_hostname` is set, so we
don't need our own code to verify the certificate.

There is no compatibility code for Python < 3.7 so the minimum version
was increased.

This also removes the unused `_match_hostname` function and its tests.