Fix for macOS ARM segfault crash in darkdetect
Pulled latest darkdetect sources from upstream to fix crash on ARM.

The reason it was crashing is because objc_msgSend has a variadic
prototype and CPython changed the way they call into such functions
in a breaking way on Python 3.9 and above. It's hard to explain but
to not have it break on ARM you need to declare the prototype properly.
:/